+++ demisto_sdk/commands/common/tests/docker_helper_test.py
@@ -1,0 +1,64 @@
+import os
+from unittest import mock
+
+import pytest
+
+import demisto_sdk.commands.common.docker_helper as dhelper
+
+
+def test_init_global_docker_client():
+    res = dhelper.init_global_docker_client(log_prompt="unit testing")
+    assert res is not None
+    assert res == dhelper.DOCKER_CLIENT
+    dhelper.DOCKER_CLIENT = None
+    # test with bad creds (should still get a valid instance)
+    with mock.patch.dict(
+        os.environ, {"DOCKERHUB_USER": "dummy", "DOCKERHUB_PASSWORD": "dummy"}
+    ):
+        res = dhelper.init_global_docker_client(log_prompt="unit testing")
+        assert res is not None
+        assert res == dhelper.DOCKER_CLIENT
+
+@pytest.mark.parametrize(
+    argnames="image, output, expected",
+    argvalues=[
+        ("alpine", b"3.7\n", "3.7"),
+        ("alpine-3", b"2.7\n", "2.7"),
+        ("alpine-310", b"3.10\n", "3.10"),
+        ("demisto/python3:3.9.8.24399", "", "3.9"),
+        ("demisto/python:2.7.18.24398", "", "2.7"),
+    ],
+)
+def test_get_python_version_from_image(
+    image: str, output: bytes, expected: float, mocker
+):
+    from demisto_sdk.commands.lint import helpers
+
+    mocker.patch.object(helpers, "init_global_docker_client")
+    helpers.init_global_docker_client().containers.run.return_value = output
+    assert expected == helpers.get_python_version_from_image(image)
+
+
+def test_cache_of_get_python_version_from_image():
+    """
+    Given -
+        docker image that should be alrady cached
+
+    When -
+        Try to get python version from am docker image
+
+    Then -
+        Validate the value returned from the cache
+    """
+    from demisto_sdk.commands.common import docker_helper
+
+    image = "demisto/python3:3.9.8.12345"
+
+    cache_info_before = docker_helper.get_python_version_from_image.cache_info()
+    docker_helper.get_python_version_from_image(image)
+    cache_info = docker_helper.get_python_version_from_image.cache_info()
+    assert cache_info.hits == cache_info_before.hits
+
+    docker_helper.get_python_version_from_image(image)
+    cache_info = docker_helper.get_python_version_from_image.cache_info()
+    assert cache_info.hits == cache_info_before.hits + 1
